requirements.txt
microdot
aioprocessing
pyserial